+++ test_autolens/point/triangles/test_extended.py
@@ -2,7 +2,7 @@ import pytest
 
 from autoarray.structures.triangles.shape import Circle
 from autolens.mock import NullTracer
-from autolens.point.solver.abstract_solver import ShapeSolver
+from autolens.point.solver.shape_solver import ShapeSolver
 
 
 @pytest.fixture
@@ -23,8 +23,6 @@ def test_solver_basic(solver):
         ),
     )
     assert list(map(tuple, result)) == [
-        (-0.012003766846269881, 0.0078125),
         (-0.0029826688901819823, -0.0078125),
-        (0.015059527021993818, -0.0078125),
-        (0.010548978043949867, 0.015625),
+        (-0.0029826688901819823, 0.0078125),
     ]
